chore: Move otel helpers into their own package
